SP_Link - add Erlang headers for Win/macOS
These need to be kept inline with headers for the specific version of Erlang the NIF is to be loaded into.

Also, headers for macOS ARM64 chipset are included but are currently not used by CMakeLists.txt - this is in anticipation of future support for Apple Silicon.

Note that the only major difference between headers on different platforms are the size declarations in erl_int_sizes_config.h